orchestration/pyproject.toml
[tool.poetry]
name = "orchestration"
version = "0.1.0"
description = "Airflow DAG authoring project"
authors = ["Platform Team <platform@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
apache-airflow = "3.1.7"


[tool.poetry.group.dev.dependencies]
black = "^26.1.0"
isort = "^8.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
